discord_bot/requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.2
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
dateparser==1.0.0
decorator==5.0.9
discord-webhook==0.14.0
idna==2.10
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
multidict==5.1.0
numpy==1.20.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
python-binance==1.0.12
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
TA-Lib==0.4.20
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.5
wcwidth==0.2.5
websockets==9.1
yarl==1.6.3
